Remove "pool control" checkbox; it implies state.
Instead put it in the drop down box and call it "policy" instead of control.  Setting pool control, especially with a checkbox, implies the pool is being given the ability to change vote bits as it sees fit.  What happens is that it just applies the current policy when the button is clicked.