Add more ticket syncing and check wallets for scripts in user db (#43)
This pull request aims to ensure that valid/invalid tickets are 
synced across all of your voting wallets.  When a user
queries /tickets it thereby asks for StakePoolUserInfo.
When this request occurs it then checks to see that
all of the wallet's lists match in length.  In the near future,
we can add more thorough checking of pool ticket statuses.

This also adds a db request about user scripts to ensure they 
are all present across the wallets.  If they aren't, then they are
added accordingly via ImportScripts along with the height
at which they registered.