Add smarter wallet connectivity
With this commit, we have aimed to properly handle the possible
disconnection of wallets' rpcclients.  Previously, when a wallet would
disconnect, rpcclient would hang and retry until a timeout would occur.
Instead with the use of DisableAutoReconnect: true in rpcclient settings
all connections are permentantly closed upon disconnect.

While this avoids timeouts it does expose the stakepool to needing to
gracefully handle the situation where some wallets are no longer available.
The new config option minservers (default: 2) offers stakepool admins
control over how many servers they require for errors not to be shown to
user in the event of disconnected wallets.  We feel 2/3 wallets
connected is sufficient for stakepool to proceed working as expected,
though anything below that may offer less than ideal situations for your
user (desync-ed votebits, address indices, and other possible corner cases
 between voting wallets).

With this new handling of wallet rpcclients, /status is now a place to
let admins see any issues with the connections to the voting wallets.
In the event of wallets going down the RPC Status: will update to
Emergency (0 wallets), Critical (1 wallet), or Degraded (2 wallets).
This can be used in conjunction with html monitoring tools, so admins
can periodically poll /status and check what the current status is of
the wallet rpcclients and notify depending on circumstances.

/status is now also used as a method to reconnect the wallets
rpcclients.  For example, wallet 1 goes down and needs to be restarted.
Admin would restart the wallet then reload /status and stakepool would
attempt to reconnect.  Each time a wallet is seen to be disconnected,
any request to /status attempts to restart the connection.  We felt like
this, though not ideal, offered admins to set the aggresiveness of
wallet reconnection as they deemed fit.

Also, an added *required* config option is adminips to limit access to
/status to a list of IPs that the admins chooses. So in the end,
users/public will not be able to access the /status page.