stakepoold: use go routines to stack network communication
Try to cram as much work as quickly as possible into the network
services. The idea here is to have as much data travel at the same point
over slow networks by doing a bit more processing up front in the
daemon. This seems to bring the overall voting latency to the sub 10ms
range.