Implement ticket ordering on tickets page.
Change the tickets data types from map to slice.
Add sorters for purchase and spent height, for live and historic tickets, respectively.
Sort live, missed, expired, and voted.  But consider not sorting live since the list may be thousands long.  Instead, this can be a start for only picking tickets with a recent spend height (would be filtered in the loop prior to sort).

Also remove dead code, stringSliceContains.