Abort stakepoold startup if gettickets fails
If stakepoold starts while dcrd RPC is inactive, it will fail to get any
tickets but will continue to start nonetheless. Therefore it won't watch
for any tickets and this will lead to missed votes.

This commit ensures that stakepoold won't silently ignore this error. It
is recommended to use a process supervisor to keep restarting stakepoold
until the RPC is ready.