controller: Never rescan during script imports
The only scripts that are imported by the controller are ones that were just
created from a createmultisig RPC, and should not have any history that requires
a rescan to find previous outputs.  Remove the rescan request from the
importscript parameters to prevent the wallet (which no longer serializes and
batches rescan jobs) to prevent extraneous resource usage for unnecessary
rescans.