Update CaptchaServe to avoid WriteImage and ServeContent.
Use captcha.NewImage, which does not have an extra copy, and privides a
Paletted image that png.(*Encoder).Encode can read from and write
directly into the http.ResponseWriter.

Avoiding http.ServeContent is also a win since no mime type detection is
needed, mod time is not relevant, and byteranges is not used.

This is rougly 2x as fast as the previous approach using a buffer,
captcha.WriteImage, and http.ServeContent.