clean up LoadTemplates and link to explorer on stats page
Previously, template.Must(httpTemplates, nil) was pointless (err=nil).
This assigns the parsed templates directly to the Templates field.

Also define a template function map for more involved template tasks.
e.g. On the stats page, use the netname function for links. This could
also be done prior to executing the template, but it may be desireable
to have the actualy network name (e.g. testnet3) still availalbe in the
Network data field.

Add debuglevel=debug to the stakepoold and dcrstakepool config used
in harness.sh.