update sample-nginx.conf
Fix the assets location. With the root directive, it would look for
files in /opt/dcrstakepool/public/assets/..., but "assets" is not a
directory that exists in the public folder.  Use the alias directive
instead to map all paths under /assets/ to the public/ folder.

Make a comment about ssl_dhparam.

Put ssl and http2 in the port 443 server's listen line.

Add the needed gzip settings.

Disable TLSv1, only using TLSv1.1 and TLSv1.2.  Also make a note about
TLSv1.3, which is in newer nginx versions.

Change the http->https rewrite to a return.  Use code 308 so that
methods are unchanged (POST stays POST).